refactor(AdminModules) AdminModules folder was created and all modules used by admin were moved into
