+++ src/pages/AdminModules/Pricings/Pricing.css
