+++ src/pages/AdminModules/Pricings/Pricing.jsx
