+++ src/pages/AdminModules/Pricings/Pricings.css
